requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
boto3==1.42.59
botocore==1.42.59
certifi==2026.2.25
click==8.3.1
distro==1.9.0
fastapi==0.134.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
jmespath==1.1.0
openai==2.24.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
structlog==25.5.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
